+++ examples/boost-asio/beast-scrapper.cpp
@@ -837,7 +837,7 @@ int main(int argc, char **argv) {
         ("timeout", po::value<std::size_t>()->default_value(5000), "generic timeout (in milliseconds)")
         ("resolve_timeout", po::value<std::size_t>()->default_value(1000), "resolve timeout (in milliseconds)")
         ("rotor_timeout", po::value<std::size_t>()->default_value(10), "rotor timeout (in milliseconds)")
-        ("max_requests", po::value<std::size_t>()->default_value(100), "maximum amount of requests before shutting down");
+        ("max_requests", po::value<std::size_t>()->default_value(20), "maximum amount of requests before shutting down");
     // clang-format on
 
     po::variables_map vm;