+++ include/rotor/asio/supervisor_asio.h
@@ -79,6 +79,9 @@ struct supervisor_asio_t : public supervisor_t {
     /** \brief returns exeuction strand */
     inline asio::io_context::strand &get_strand() noexcept { return *strand; }
 
+    /** \brief process queue of messages of locality leader */
+    void do_process() noexcept;
+
   protected:
     /** \struct timer_t
      * \brief boos::asio::deadline_timer with embedded timer handler */