+++ src/rotor/plugin/delivery.cpp
@@ -43,7 +43,7 @@ std::string inspected_local_delivery_t::identify(const message_base_t *message, 
     using boost::core::demangle;
     using T = owner_tag_t;
     std::string info = demangle((const char *)message->type_index);
-    std::int32_t level = 0;
+    std::int32_t level = 5;
     auto dump_point = [](const subscription_point_t &p) -> std::string {
         std::stringstream out;
         out << " [";
@@ -82,9 +82,67 @@ std::string inspected_local_delivery_t::identify(const message_base_t *message, 
         level = 9;
         info += dump_point(m->payload.point);
     } else if (auto m = dynamic_cast<const message::deregistration_service_t *>(message); m) {
-        level = 0;
+        level = 1;
         info += ", service = ";
         info += m->payload.service_name;
+    } else if (auto m = dynamic_cast<const message::registration_request_t *>(message); m) {
+        level = 1;
+        info += ", name = ";
+        info += m->payload.request_payload.service_name;
+    } else if (auto m = dynamic_cast<const message::registration_response_t *>(message); m) {
+        level = 1;
+        info += ", name = ";
+        info += m->payload.req->payload.request_payload.service_name;
+        auto &ee = m->payload.ee;
+        if (ee) {
+            info += ", failure(!) = ";
+            info += ee->message() + " ";
+        } else {
+            info += ", success; ";
+        }
+    } else if (auto m = dynamic_cast<const message::discovery_request_t *>(message); m) {
+        level = 1;
+        info += ", service = ";
+        info += m->payload.request_payload.service_name;
+    } else if (auto m = dynamic_cast<const message::discovery_promise_t *>(message); m) {
+        level = 1;
+        info += ", service = ";
+        info += m->payload.request_payload.service_name;
+    } else if (auto m = dynamic_cast<const message::discovery_future_t *>(message); m) {
+        level = 1;
+        std::stringstream out;
+        out << ", service = " << m->payload.req->payload.request_payload.service_name;
+        auto &ee = m->payload.ee;
+        if (ee) {
+            out << ", failure = " << ee->message() << " ";
+        } else {
+            auto &addr = m->payload.res.service_addr;
+            out << ", success, addr =  " << addr.get() << " ";
+        }
+        info += out.str();
+    } else if (auto m = dynamic_cast<const message::discovery_response_t *>(message); m) {
+        level = 1;
+        std::stringstream out;
+        out << ", service = " << m->payload.req->payload.request_payload.service_name;
+        auto &ee = m->payload.ee;
+        if (ee) {
+            out << ", failure = " << ee->message() << " ";
+        } else {
+            auto &addr = m->payload.res.service_addr;
+            out << ", success, addr =  " << addr.get() << " ";
+        }
+        info += out.str();
+    } else if (auto m = dynamic_cast<const message::link_response_t *>(message); m) {
+        level = 2;
+        std::stringstream out;
+        out << ", service addr = " << m->payload.req->payload.origin.get();
+        auto &ee = m->payload.ee;
+        if (ee) {
+            out << ", failure(!) = " << ee->message() + " ";
+        } else {
+            out << ", success; ";
+        }
+        info += out.str();
     } else if (auto m = dynamic_cast<const message::shutdown_trigger_t *>(message); m) {
         level = 0;
         std::stringstream out;