+++ src/rotor/thread/supervisor_thread.cpp
@@ -18,13 +18,14 @@ void supervisor_thread_t::shutdown() noexcept {
     auto &sup_addr = supervisor->get_address();
     auto ec = make_error_code(shutdown_code_t::normal);
     auto reason = make_error(ec);
-    supervisor->enqueue(make_message<payload::shutdown_trigger_t>(sup_addr, address, reason));
+    auto msg = make_message<payload::shutdown_trigger_t>(sup_addr, address, reason);
+    supervisor->enqueue(msg);
 }
 
 void supervisor_thread_t::enqueue(message_ptr_t message) noexcept {
     auto ctx = static_cast<system_context_thread_t *>(context);
+    inbound_queue.push(message.detach());
     std::lock_guard<std::mutex> lock(ctx->mutex);
-    ctx->inbound.emplace_back(std::move(message));
     ctx->cv.notify_one();
 }